+++ launcher.spec
@@ -90,7 +90,7 @@ hiddenimports = (
         "psygnal._weak_callback",
         "imagecodecs._imagecodecs",
         "PartSeg.plugins.napari_widgets",
-        "PartSegCore.napari_plugins",
+        "PartSegCore.napari_io",
     ]
     + [x.module_name for x in imageio_known_plugins.values()]
     + [x for x in collect_submodules("skimage") if "tests" not in x]
