+++ package/PartSeg/plugins/napari_io/load_image.py
@@ -1,7 +1,7 @@
 import functools
 
+from PartSeg.plugins.napari_io.loader import partseg_loader
 from PartSegCore.analysis.load_functions import LoadStackImage
-from PartSegCore.napari_plugins.loader import partseg_loader
 
 
 def napari_get_reader(path: str):
