+++ package/PartSeg/plugins/napari_io/load_mask_project.py
@@ -1,7 +1,7 @@
 import functools
 
+from PartSeg.plugins.napari_io.loader import partseg_loader
 from PartSegCore.mask.io_functions import LoadROI
-from PartSegCore.napari_plugins.loader import partseg_loader
 
 
 def napari_get_reader(path: str):
